多层次父类提取
