+++ src/main/java/com/github/hcsp/inheritance/Girl.java
@@ -1,21 +1,12 @@
 package com.github.hcsp.inheritance;
 
-public class Girl {
-    private String name;
-    private int age;
+public class Girl extends  AnimalPeople{
 
     public Girl(String name, int age) {
-        this.name = name;
-        this.age = age;
+        super(name, age);
     }
 
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
-    }
 
-    public void sayMyAge() {
-        System.out.println("我的年龄是" + age);
-    }
 
     public void sayGirl() {
         System.out.println("我是一个女孩");
